refactor(builtins): split arith.rs into cohesive submodules (§7-8)
Split the 1528-line builtins/arith.rs (over the 500-line limit) into a
facade plus six themed submodules under src/builtins/arith/. Every
submodule is under 500 lines. The facade re-exports all externally-used
symbols at their original visibility so no external caller changes.

- arith.rs (23)         - facade: mod decls + named re-exports
- rat.rs (211)          - Rat/BigNum helpers (rat_*_checked, is_fat_rat_like,
                          bigint_ratio_to_f64, float_mod_floor, as_bigint,
                          to_big_rat_parts, make_fat_rat, real_to_rat)
- temporal.rs (126)     - Date/Instant/Duration helpers (is_temporal_operand,
                          instance_days/instant/duration/datetime, make_duration*)
- range.rs (238)        - Range arithmetic (range_bounds/offset/scale/divide,
                          canonical_int_range, mixin_range_arith*)
- add_sub.rs (345)      - arith_add, arith_add_coerced, arith_sub
- mul_div_mod.rs (346)  - arith_mul, arith_div, arith_mod
- pow_negate.rs (319)   - arith_pow, arith_negate

External callers verified unchanged (no edits required):
  src/builtins/mod.rs           - uses arith::{arith_add,sub,mul,div,mod,pow,negate}
  src/value/value_lazy.rs       - uses crate::builtins::arith::{arith_*}
  src/vm/vm_arith_ops.rs        - uses crate::builtins::arith::is_temporal_operand
  src/runtime/builtins_system_async.rs - uses crate::builtins::arith::make_duration_value
  src/runtime/methods_object_native_ctors_temporal.rs - uses crate::builtins::arith::real_to_rat

Pure relocation. Moved fns are pub(crate) in submodules; cross-submodule
references use super:: imports. The fn multiset is byte-identical.
cargo build + clippy -D warnings clean; make test 12703/12703 PASS.
