+++ src/builtins/arith.rs
@@ -1,1528 +1,23 @@
 #![allow(clippy::result_large_err)]
+//! Arithmetic operator implementations: facade re-exporting themed submodules.
+//!
+//! Split from the original monolithic `arith.rs`. Each submodule owns a
+//! cohesive group of free functions; this facade preserves every public
+//! function's original visibility through explicit named re-exports so
+//! external callers (`arith::<name>`) keep working unchanged.
 
-use std::sync::Arc;
+mod add_sub;
+mod mul_div_mod;
+mod pow_negate;
+mod range;
+mod rat;
+mod temporal;
 
-use crate::runtime;
-use crate::symbol::Symbol;
-use crate::value::{RuntimeError, Value, make_big_fat_rat, make_big_rat_arith, make_rat};
-use num_bigint::{BigInt as NumBigInt, Sign};
-use num_traits::{Signed, ToPrimitive, Zero};
+// Core arithmetic operators — re-exported pub(crate) to match original visibility.
+pub(crate) use add_sub::{arith_add, arith_sub};
+pub(crate) use mul_div_mod::{arith_div, arith_mod, arith_mul};
+pub(crate) use pow_negate::{arith_negate, arith_pow};
 
-/// Perform Rat addition with overflow detection.
-/// Returns Num if the result's denominator would exceed i64 range.
-fn rat_add_checked(an: i64, ad: i64, bn: i64, bd: i64) -> Value {
-    // Use i128 to detect overflow
-    let n = an as i128 * bd as i128 + bn as i128 * ad as i128;
-    let d = ad as i128 * bd as i128;
-    rat_from_i128_or_num(n, d)
-}
-
-/// Perform Rat subtraction with overflow detection.
-fn rat_sub_checked(an: i64, ad: i64, bn: i64, bd: i64) -> Value {
-    let n = an as i128 * bd as i128 - bn as i128 * ad as i128;
-    let d = ad as i128 * bd as i128;
-    rat_from_i128_or_num(n, d)
-}
-
-/// Perform Rat multiplication with overflow detection.
-fn rat_mul_checked(an: i64, ad: i64, bn: i64, bd: i64) -> Value {
-    let n = an as i128 * bn as i128;
-    let d = ad as i128 * bd as i128;
-    rat_from_i128_or_num(n, d)
-}
-
-/// Perform Rat division with overflow detection.
-fn rat_div_checked(an: i64, ad: i64, bn: i64, bd: i64) -> Value {
-    let n = an as i128 * bd as i128;
-    let d = ad as i128 * bn as i128;
-    rat_from_i128_or_num(n, d)
-}
-
-/// Convert i128 numerator/denominator to Rat, or Num on overflow.
-/// Raku spec: Rat denominators are limited to uint64 range.
-/// When the denominator exceeds this after GCD reduction, degrade to Num.
-fn rat_from_i128_or_num(n: i128, d: i128) -> Value {
-    if d == 0 {
-        return if n == 0 {
-            Value::Rat(0, 0)
-        } else if n > 0 {
-            Value::Rat(1, 0)
-        } else {
-            Value::Rat(-1, 0)
-        };
-    }
-    // Normalize sign
-    let (mut n, mut d) = if d < 0 { (-n, -d) } else { (n, d) };
-    // GCD
-    fn gcd128(mut a: i128, mut b: i128) -> i128 {
-        a = a.abs();
-        b = b.abs();
-        while b != 0 {
-            let t = b;
-            b = a % b;
-            a = t;
-        }
-        a
-    }
-    let g = gcd128(n, d);
-    if g != 0 {
-        n /= g;
-        d /= g;
-    }
-    // Check if result fits in i64 (denominator must also fit in uint64 per Raku spec)
-    if let (Ok(n64), Ok(d64)) = (i64::try_from(n), i64::try_from(d)) {
-        return Value::Rat(n64, d64);
-    }
-    // Overflow: degrade to Num
-    Value::Num(n as f64 / d as f64)
-}
-
-/// Check if BigRat arithmetic is needed: at least one operand requires BigInt precision
-/// (BigRat or BigInt) AND at least one operand is rational (Rat/FatRat/BigRat).
-/// Plain Rat-vs-Rat operations use the i128-based overflow-to-Num path instead.
-fn needs_bigrat_path(l: &Value, r: &Value) -> bool {
-    let has_big = matches!(l, Value::BigRat(_, _) | Value::BigInt(_))
-        || matches!(r, Value::BigRat(_, _) | Value::BigInt(_));
-    let has_rat = matches!(
-        l,
-        Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-    ) || matches!(
-        r,
-        Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-    );
-    has_big && has_rat
-}
-
-/// Check if a value should be treated as FatRat for arithmetic.
-/// This includes Value::FatRat and Value::BigRat when the denominator exceeds u64,
-/// since only FatRat operations can produce such values (regular Rat degrades to Num).
-fn is_fat_rat_like(v: &Value) -> bool {
-    match v {
-        Value::FatRat(_, _) => true,
-        Value::BigRat(_, d) => d.to_u64().is_none(),
-        _ => false,
-    }
-}
-
-/// Safely convert a BigInt ratio to f64, handling cases where both
-/// numerator and denominator are too large for f64 individually.
-fn bigint_ratio_to_f64(n: &NumBigInt, d: &NumBigInt) -> f64 {
-    if d.is_zero() {
-        if n.is_zero() {
-            return f64::NAN;
-        }
-        return if n.is_negative() {
-            f64::NEG_INFINITY
-        } else {
-            f64::INFINITY
-        };
-    }
-    // Try direct conversion first
-    if let (Some(nf), Some(df)) = (n.to_f64(), d.to_f64())
-        && nf.is_finite()
-        && df.is_finite()
-        && df != 0.0
-    {
-        return nf / df;
-    }
-    // For very large numbers, use scaled division
-    let sign = n.is_negative() ^ d.is_negative();
-    let na = n.abs();
-    let da = d.abs();
-    // Scale to get enough precision
-    let scaled = &na * NumBigInt::from(10u64).pow(20);
-    let div = &scaled / &da;
-    let val = div.to_f64().unwrap_or(0.0) / 1e20;
-    if sign { -val } else { val }
-}
-
-/// Floored modulo for f64: result has the same sign as the divisor.
-fn float_mod_floor(a: f64, b: f64) -> f64 {
-    let r = a % b;
-    if (r > 0.0 && b < 0.0) || (r < 0.0 && b > 0.0) {
-        r + b
-    } else {
-        r
-    }
-}
-
-/// Check if a value is a Date, Instant, or Duration instance (temporal operand for arithmetic).
-pub fn is_temporal_operand(value: &Value) -> bool {
-    matches!(value, Value::Instance { class_name, .. }
-        if class_name == "Date" || class_name == "Instant" || class_name == "Duration")
-}
-
-fn as_bigint(value: &Value) -> Option<NumBigInt> {
-    match value {
-        Value::Int(i) => Some(NumBigInt::from(*i)),
-        Value::BigInt(i) => Some((**i).clone()),
-        _ => None,
-    }
-}
-
-fn to_big_rat_parts(value: &Value) -> Option<(NumBigInt, NumBigInt)> {
-    match value {
-        Value::Int(i) => Some((NumBigInt::from(*i), NumBigInt::from(1))),
-        Value::BigInt(i) => Some(((**i).clone(), NumBigInt::from(1))),
-        Value::Rat(n, d) | Value::FatRat(n, d) => Some((NumBigInt::from(*n), NumBigInt::from(*d))),
-        Value::BigRat(n, d) => Some(((**n).clone(), (**d).clone())),
-        _ => None,
-    }
-}
-
-fn instance_days(value: &Value) -> Option<i64> {
-    match value {
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Date" => match attributes.as_map().get("days") {
-            Some(Value::Int(days)) => Some(*days),
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-fn instance_instant_value(value: &Value) -> Option<f64> {
-    match value {
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Instant" => attributes
-            .as_map()
-            .get("value")
-            .and_then(runtime::to_float_value),
-        _ => None,
-    }
-}
-fn instance_instant_raw(value: &Value) -> Option<Value> {
-    match value {
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Instant" => attributes.as_map().get("value").cloned(),
-        _ => None,
-    }
-}
-fn value_sub(a: Value, b: Value) -> Value {
-    let (l, r) = runtime::coerce_numeric(a, b);
-    if let (Some((an, ad)), Some((bn, bd))) = (to_big_rat_parts(&l), to_big_rat_parts(&r)) {
-        return crate::value::make_big_rat_arith(an * &bd - bn * &ad, ad * bd);
-    }
-    Value::Num(
-        runtime::to_float_value(&l).unwrap_or(0.0) - runtime::to_float_value(&r).unwrap_or(0.0),
-    )
-}
-
-fn instance_duration_value(value: &Value) -> Option<f64> {
-    match value {
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Duration" => attributes
-            .as_map()
-            .get("value")
-            .and_then(runtime::to_float_value),
-        _ => None,
-    }
-}
-
-/// Return the raw stored `value` of a Duration instance (a Rational), if any.
-fn instance_duration_raw_value(value: &Value) -> Option<Value> {
-    match value {
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Duration" => attributes.as_map().get("value").cloned(),
-        _ => None,
-    }
-}
-
-/// Coerce a Real value to a Rational, matching Rakudo's Duration which always
-/// stores its seconds as a Rat. Integers become exact Rats; existing rationals
-/// are kept; Inf/NaN map to the degenerate Rats `1/0`/`0/0`; floats use the
-/// standard epsilon conversion.
-pub(crate) fn real_to_rat(v: &Value) -> Value {
-    match v {
-        Value::Rat(..) | Value::BigRat(..) | Value::FatRat(..) => v.clone(),
-        Value::Int(i) => crate::value::make_rat(*i, 1),
-        Value::Bool(b) => crate::value::make_rat(*b as i64, 1),
-        // Preserve large integers exactly as a (big) Rat with denominator 1.
-        Value::BigInt(b) => crate::value::make_big_rat_arith((**b).clone(), NumBigInt::from(1)),
-        Value::Num(f) => {
-            if f.is_nan() {
-                Value::Rat(0, 0)
-            } else if f.is_infinite() {
-                if *f > 0.0 {
-                    Value::Rat(1, 0)
-                } else {
-                    Value::Rat(-1, 0)
-                }
-            } else {
-                crate::builtins::num_to_rat_with_epsilon(*f, 1e-6)
-            }
-        }
-        // Anything else (e.g. Str, allomorphs): fall back through f64.
-        _ => {
-            let f = runtime::to_float_value(v).unwrap_or(0.0);
-            if f.is_nan() {
-                Value::Rat(0, 0)
-            } else if f.is_infinite() {
-                if f > 0.0 {
-                    Value::Rat(1, 0)
-                } else {
-                    Value::Rat(-1, 0)
-                }
-            } else {
-                crate::builtins::num_to_rat_with_epsilon(f, 1e-6)
-            }
-        }
-    }
-}
-
-/// Build a Duration instance storing the given (already Rational) value.
-fn make_duration_from_value(val: Value) -> Value {
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("value".to_string(), val);
-    Value::make_instance(Symbol::intern("Duration"), attrs)
-}
-
-fn instance_datetime_parts(value: &Value) -> Option<(i64, i64, i64, i64, i64, f64, i64)> {
-    match value {
-        Value::Instance { attributes, .. }
-            if attributes.contains_key("year")
-                && attributes.contains_key("month")
-                && attributes.contains_key("day")
-                && attributes.contains_key("hour")
-                && attributes.contains_key("minute")
-                && attributes.contains_key("second")
-                && attributes.contains_key("timezone") =>
-        {
-            Some(crate::builtins::methods_0arg::temporal::datetime_attrs(
-                &(attributes).as_map(),
-            ))
-        }
-        _ => None,
-    }
-}
-
-pub fn make_duration_value(secs: f64) -> Value {
-    make_duration(secs)
-}
-
-fn make_duration(secs: f64) -> Value {
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("value".to_string(), Value::Num(secs));
-    Value::make_instance(Symbol::intern("Duration"), attrs)
-}
-
-fn make_fat_rat(num: i64, den: i64) -> Value {
-    match make_rat(num, den) {
-        Value::Rat(n, d) => Value::FatRat(n, d),
-        other => other,
-    }
-}
-
-/// Scale a Range by a numeric factor. Returns Some(range) if `range_val` is a Range
-/// and `factor` is numeric, None otherwise.
-/// Extract `(start, end, excl_start, excl_end)` from any Range variant
-/// (the int-endpoint variants or a `GenericRange`). Returns `None` for
-/// non-Range values.
-fn range_bounds(range_val: &Value) -> Option<(Value, Value, bool, bool)> {
-    Some(match range_val {
-        Value::Range(a, b) => (Value::Int(*a), Value::Int(*b), false, false),
-        Value::RangeExcl(a, b) => (Value::Int(*a), Value::Int(*b), false, true),
-        Value::RangeExclStart(a, b) => (Value::Int(*a), Value::Int(*b), true, false),
-        Value::RangeExclBoth(a, b) => (Value::Int(*a), Value::Int(*b), true, true),
-        Value::GenericRange {
-            start,
-            end,
-            excl_start,
-            excl_end,
-        } => (
-            start.as_ref().clone(),
-            end.as_ref().clone(),
-            *excl_start,
-            *excl_end,
-        ),
-        _ => return None,
-    })
-}
-
-/// Offset both bounds of a Range by `delta` using `op` (add or subtract),
-/// preserving exclusivity. Returns `Some(range)` when `range_val` is a Range and
-/// `delta` is a real number (Int/Num/Rat/FatRat/BigInt). Collapses to a
-/// canonical integer range when both endpoints remain `Int` (so `(2..4) + 1`
-/// is `eqv` to the literal `3..5`); falls back to `GenericRange` otherwise.
-fn range_offset(
-    range_val: &Value,
-    delta: &Value,
-    op: impl Fn(Value, Value) -> Value,
-) -> Option<Value> {
-    if !matches!(
-        delta,
-        Value::Int(_) | Value::Num(_) | Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigInt(_)
-    ) {
-        return None;
-    }
-    let (start, end, excl_start, excl_end) = range_bounds(range_val)?;
-    Some(canonical_int_range(
-        op(start, delta.clone()),
-        op(end, delta.clone()),
-        excl_start,
-        excl_end,
-    ))
-}
-
-fn range_scale(range_val: &Value, factor: &Value) -> Option<Value> {
-    let n = match factor {
-        Value::Int(i) => Some(*i as f64),
-        Value::Num(f) => Some(*f),
-        Value::Rat(n, d) if *d != 0 => Some(*n as f64 / *d as f64),
-        _ => None,
-    }?;
-    let (start, end, excl_start, excl_end) = range_bounds(range_val)?;
-    let scale_val = |v: Value| -> Value {
-        match v {
-            Value::Int(i) => {
-                let result = i as f64 * n;
-                if result == result.floor() && result.abs() < i64::MAX as f64 {
-                    Value::Int(result as i64)
-                } else {
-                    Value::Num(result)
-                }
-            }
-            Value::Num(f) => Value::Num(f * n),
-            Value::Rat(rn, rd) if rd != 0 => Value::Num(rn as f64 / rd as f64 * n),
-            other => Value::Num(other.to_f64() * n),
-        }
-    };
-    let new_start = scale_val(start);
-    let new_end = scale_val(end);
-    Some(canonical_int_range(
-        new_start, new_end, excl_start, excl_end,
-    ))
-}
-
-/// Build a Range value, collapsing to the canonical integer-range variant
-/// (`Range`/`RangeExcl`/`RangeExclStart`/`RangeExclBoth`) when both endpoints
-/// are `Int`, so the result is `eqv` to a literal range (e.g. `(2..^5) * 5`
-/// must equal `10..^25`). Falls back to `GenericRange` for non-Int endpoints.
-fn canonical_int_range(start: Value, end: Value, excl_start: bool, excl_end: bool) -> Value {
-    if let (Value::Int(a), Value::Int(b)) = (&start, &end) {
-        let (a, b) = (*a, *b);
-        return match (excl_start, excl_end) {
-            (false, false) => Value::Range(a, b),
-            (false, true) => Value::RangeExcl(a, b),
-            (true, false) => Value::RangeExclStart(a, b),
-            (true, true) => Value::RangeExclBoth(a, b),
-        };
-    }
-    Value::GenericRange {
-        start: Arc::new(start),
-        end: Arc::new(end),
-        excl_start,
-        excl_end,
-    }
-}
-
-/// Divide a Range by a numeric factor. Returns Some(range) if the left is a Range.
-fn range_divide(range_val: &Value, divisor: &Value) -> Option<Value> {
-    // Get divisor as rational (num, den) for exact division
-    let (div_n, div_d): (i64, i64) = match divisor {
-        Value::Int(i) if *i != 0 => (*i, 1),
-        Value::Rat(n, d) if *d != 0 && *n != 0 => (*n, *d),
-        Value::Num(f) if *f != 0.0 => return range_divide_float(range_val, *f),
-        _ => return None,
-    };
-    let (start, end, excl_start, excl_end) = match range_val {
-        Value::Range(a, b) => (Value::Int(*a), Value::Int(*b), false, false),
-        Value::RangeExcl(a, b) => (Value::Int(*a), Value::Int(*b), false, true),
-        Value::RangeExclStart(a, b) => (Value::Int(*a), Value::Int(*b), true, false),
-        Value::RangeExclBoth(a, b) => (Value::Int(*a), Value::Int(*b), true, true),
-        Value::GenericRange {
-            start,
-            end,
-            excl_start,
-            excl_end,
-        } => (
-            start.as_ref().clone(),
-            end.as_ref().clone(),
-            *excl_start,
-            *excl_end,
-        ),
-        _ => return None,
-    };
-    // Dividing by (div_n/div_d) = multiplying by (div_d/div_n)
-    // Use make_rat for proper normalization, but preserve Rat type (don't collapse to Int)
-    let div_val = |v: Value| -> Value {
-        let num = match &v {
-            Value::Int(i) => *i,
-            Value::Rat(n, d) => {
-                // (n/d) / (div_n/div_d) = (n*div_d) / (d*div_n)
-                let result = make_rat(n * div_d, d * div_n);
-                // Ensure result stays Rat even if whole number
-                if let Value::Int(i) = result {
-                    return Value::Rat(i, 1);
-                }
-                return result;
-            }
-            _ => return Value::Num(v.to_f64() * div_d as f64 / div_n as f64),
-        };
-        let result = make_rat(num * div_d, div_n);
-        // Ensure result stays Rat even if whole number
-        if let Value::Int(i) = result {
-            return Value::Rat(i, 1);
-        }
-        result
-    };
-    let new_start = div_val(start);
-    let new_end = div_val(end);
-    Some(Value::GenericRange {
-        start: Arc::new(new_start),
-        end: Arc::new(new_end),
-        excl_start,
-        excl_end,
-    })
-}
-
-/// Divide a Range by a float factor.
-fn range_divide_float(range_val: &Value, n: f64) -> Option<Value> {
-    let (start, end, excl_start, excl_end) = match range_val {
-        Value::Range(a, b) => (Value::Int(*a), Value::Int(*b), false, false),
-        Value::RangeExcl(a, b) => (Value::Int(*a), Value::Int(*b), false, true),
-        Value::RangeExclStart(a, b) => (Value::Int(*a), Value::Int(*b), true, false),
-        Value::RangeExclBoth(a, b) => (Value::Int(*a), Value::Int(*b), true, true),
-        Value::GenericRange {
-            start,
-            end,
-            excl_start,
-            excl_end,
-        } => (
-            start.as_ref().clone(),
-            end.as_ref().clone(),
-            *excl_start,
-            *excl_end,
-        ),
-        _ => return None,
-    };
-    let div_val = |v: Value| -> Value { Value::Num(v.to_f64() / n) };
-    let new_start = div_val(start);
-    let new_end = div_val(end);
-    Some(Value::GenericRange {
-        start: Arc::new(new_start),
-        end: Arc::new(new_end),
-        excl_start,
-        excl_end,
-    })
-}
-
-// ── Mixin-Range arithmetic helper ────────────────────────────────────
-/// If the left operand is a Mixin wrapping a Range (e.g. `(2..^5) but Meows`),
-/// perform the arithmetic on the inner Range and re-wrap the result with the
-/// same Mixin. Returns `None` if neither operand is a Mixin-wrapped Range.
-fn mixin_range_arith<F>(left: Value, right: Value, op: F) -> Option<Result<Value, RuntimeError>>
-where
-    F: Fn(Value, Value) -> Result<Value, RuntimeError>,
-{
-    match &left {
-        Value::Mixin(inner, mixins) if inner.is_range() => {
-            let inner_val = (**inner).clone();
-            let mix_clone = (**mixins).clone();
-            Some(op(inner_val, right).map(|r| Value::mixin(r, mix_clone)))
-        }
-        _ => None,
-    }
-}
-
-/// Same as `mixin_range_arith` but for non-Result arithmetic functions.
-fn mixin_range_arith_val<F>(left: Value, right: Value, op: F) -> Option<Value>
-where
-    F: Fn(Value, Value) -> Value,
-{
-    match &left {
-        Value::Mixin(inner, mixins) if inner.is_range() => {
-            let result = op((**inner).clone(), right);
-            Some(Value::mixin(result, (**mixins).clone()))
-        }
-        _ => None,
-    }
-}
-
-// ── Arithmetic operators ─────────────────────────────────────────────
-pub(crate) fn arith_add(left: Value, right: Value) -> Result<Value, RuntimeError> {
-    // Phase 2 element container: a `:=`-bound element cell may reach an arith op
-    // directly (e.g. `@a.reduce(&[+])` folds raw items); read through the cell.
-    let (left, right) = (left.into_deref(), right.into_deref());
-    // Mixin-wrapped Range + Real (or Real + Mixin Range): perform Range arithmetic and re-wrap
-    if let Some(result) = mixin_range_arith(left.clone(), right.clone(), arith_add)
-        .or_else(|| mixin_range_arith(right.clone(), left.clone(), arith_add))
-    {
-        return result;
-    }
-    // Range + Real (commutative): shift both bounds, preserving exclusivity.
-    let add_endpoint = |a: Value, b: Value| arith_add(a, b).unwrap_or(Value::Int(0));
-    if let Some(range) = range_offset(&left, &right, add_endpoint)
-        .or_else(|| range_offset(&right, &left, add_endpoint))
-    {
-        return Ok(range);
-    }
-    // Date + Int: add days
-    if let Some(days) = instance_days(&left)
-        && let Value::Int(delta) = &right
-    {
-        use crate::builtins::methods_0arg::temporal;
-        let new_days = days + delta;
-        let (y, m, d) = temporal::epoch_days_to_civil(new_days);
-        return Ok(temporal::make_date(y, m, d));
-    }
-    if let Some(days) = instance_days(&right)
-        && let Value::Int(delta) = &left
-    {
-        use crate::builtins::methods_0arg::temporal;
-        let new_days = days + delta;
-        let (y, m, d) = temporal::epoch_days_to_civil(new_days);
-        return Ok(temporal::make_date(y, m, d));
-    }
-    // Instant + Instant is illegal
-    if instance_instant_value(&left).is_some() && instance_instant_value(&right).is_some() {
-        return Err(RuntimeError::new(
-            "Cannot add two Instants together".to_string(),
-        ));
-    }
-    // Instant + Duration => Instant
-    if let Some(tai) = instance_instant_value(&left)
-        && let Some(dur) = instance_duration_value(&right)
-    {
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("value".to_string(), Value::Num(tai + dur));
-        return Ok(Value::make_instance(Symbol::intern("Instant"), attrs));
-    }
-    if let Some(tai) = instance_instant_value(&right)
-        && let Some(dur) = instance_duration_value(&left)
-    {
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("value".to_string(), Value::Num(tai + dur));
-        return Ok(Value::make_instance(Symbol::intern("Instant"), attrs));
-    }
-    // Instant + Numeric => Instant (add to TAI value)
-    if let Some(tai) = instance_instant_value(&left)
-        && right.is_numeric()
-    {
-        let delta = runtime::to_float_value(&right).unwrap_or(0.0);
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("value".to_string(), Value::Num(tai + delta));
-        return Ok(Value::make_instance(Symbol::intern("Instant"), attrs));
-    }
-    if let Some(tai) = instance_instant_value(&right)
-        && left.is_numeric()
-    {
-        let delta = runtime::to_float_value(&left).unwrap_or(0.0);
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("value".to_string(), Value::Num(tai + delta));
-        return Ok(Value::make_instance(Symbol::intern("Instant"), attrs));
-    }
-    // Duration + Numeric => Duration
-    if let Some(dur) = instance_duration_value(&left)
-        && right.is_numeric()
-    {
-        let delta = runtime::to_float_value(&right).unwrap_or(0.0);
-        return Ok(make_duration(dur + delta));
-    }
-    if let Some(dur) = instance_duration_value(&right)
-        && left.is_numeric()
-    {
-        let delta = runtime::to_float_value(&left).unwrap_or(0.0);
-        return Ok(make_duration(dur + delta));
-    }
-    // DateTime + Duration => DateTime
-    if let Some((y, m, d, h, mi, s, tz)) = instance_datetime_parts(&left)
-        && let Some(delta) = instance_duration_value(&right)
-    {
-        use crate::builtins::methods_0arg::temporal;
-        let instant = temporal::datetime_to_instant_leap_aware(y, m, d, h, mi, s, tz);
-        let (ny, nm, nd, nh, nmi, ns) =
-            temporal::instant_to_datetime_leap_aware(instant + delta, tz);
-        return Ok(temporal::make_datetime(ny, nm, nd, nh, nmi, ns, tz));
-    }
-    if let Some(delta) = instance_duration_value(&left)
-        && let Some((y, m, d, h, mi, s, tz)) = instance_datetime_parts(&right)
-    {
-        use crate::builtins::methods_0arg::temporal;
-        let instant = temporal::datetime_to_instant_leap_aware(y, m, d, h, mi, s, tz);
-        let (ny, nm, nd, nh, nmi, ns) =
-            temporal::instant_to_datetime_leap_aware(instant + delta, tz);
-        return Ok(temporal::make_datetime(ny, nm, nd, nh, nmi, ns, tz));
-    }
-    let (l, r) = runtime::coerce_numeric(left, right);
-    Ok(arith_add_coerced(l, r))
-}
-
-fn arith_add_coerced(l: Value, r: Value) -> Value {
-    if matches!(l, Value::Complex(_, _)) || matches!(r, Value::Complex(_, _)) {
-        let (ar, ai) = runtime::to_complex_parts(&l).unwrap_or((0.0, 0.0));
-        let (br, bi) = runtime::to_complex_parts(&r).unwrap_or((0.0, 0.0));
-        Value::Complex(ar + br, ai + bi)
-    } else if (matches!(l, Value::BigInt(_)) || matches!(r, Value::BigInt(_)))
-        && let (Some(a), Some(b)) = (as_bigint(&l), as_bigint(&r))
-    {
-        Value::from_bigint(a + b)
-    } else if let (Some((an, ad)), Some((bn, bd))) = (to_big_rat_parts(&l), to_big_rat_parts(&r))
-        && needs_bigrat_path(&l, &r)
-    {
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        if has_fat_rat {
-            match make_big_fat_rat(an * bd.clone() + bn * ad.clone(), ad * bd) {
-                Value::Rat(n, d) => Value::FatRat(n, d),
-                other => other,
-            }
-        } else {
-            make_big_rat_arith(an * bd.clone() + bn * ad.clone(), ad * bd)
-        }
-    } else if let (Some((an, ad)), Some((bn, bd))) =
-        (runtime::to_rat_parts(&l), runtime::to_rat_parts(&r))
-    {
-        let has_rat = matches!(l, Value::Rat(_, _) | Value::FatRat(_, _))
-            || matches!(r, Value::Rat(_, _) | Value::FatRat(_, _));
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        if has_rat {
-            if has_fat_rat {
-                let n = an * bd + bn * ad;
-                let d = ad * bd;
-                make_fat_rat(n, d)
-            } else {
-                rat_add_checked(an, ad, bn, bd)
-            }
-        } else {
-            match (l, r) {
-                (Value::Int(a), Value::Int(b)) => match a.checked_add(b) {
-                    Some(sum) => Value::Int(sum),
-                    None => Value::from_bigint(
-                        num_bigint::BigInt::from(a) + num_bigint::BigInt::from(b),
-                    ),
-                },
-                (Value::Num(a), Value::Num(b)) => Value::Num(a + b),
-                (Value::Int(a), Value::Num(b)) => Value::Num(a as f64 + b),
-                (Value::Num(a), Value::Int(b)) => Value::Num(a + b as f64),
-                _ => Value::Int(0),
-            }
-        }
-    } else {
-        let lf = runtime::to_float_value(&l);
-        let rf = runtime::to_float_value(&r);
-        if let (Some(a), Some(b)) = (lf, rf) {
-            Value::Num(a + b)
-        } else {
-            match (l, r) {
-                (Value::Int(a), Value::Int(b)) => match a.checked_add(b) {
-                    Some(sum) => Value::Int(sum),
-                    None => Value::from_bigint(
-                        num_bigint::BigInt::from(a) + num_bigint::BigInt::from(b),
-                    ),
-                },
-                (Value::Num(a), Value::Num(b)) => Value::Num(a + b),
-                (Value::Int(a), Value::Num(b)) => Value::Num(a as f64 + b),
-                (Value::Num(a), Value::Int(b)) => Value::Num(a + b as f64),
-                _ => Value::Int(0),
-            }
-        }
-    }
-}
-
-pub(crate) fn arith_sub(left: Value, right: Value) -> Value {
-    let (left, right) = (left.into_deref(), right.into_deref());
-    if let (Some(a), Some(b)) = (instance_instant_raw(&left), instance_instant_raw(&right)) {
-        return make_duration(runtime::to_float_value(&value_sub(a, b)).unwrap_or(0.0));
-    }
-    if let Some(a) = instance_instant_raw(&left)
-        && let Some(dur) = instance_duration_value(&right)
-    {
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("value".to_string(), value_sub(a, Value::Num(dur)));
-        return Value::make_instance(Symbol::intern("Instant"), attrs);
-    }
-    if let Some(a) = instance_instant_raw(&left)
-        && right.is_numeric()
-    {
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("value".to_string(), value_sub(a, right));
-        return Value::make_instance(Symbol::intern("Instant"), attrs);
-    }
-    // Duration - Duration returns Duration
-    if let (Some(a), Some(b)) = (
-        instance_duration_value(&left),
-        instance_duration_value(&right),
-    ) {
-        return make_duration(a - b);
-    }
-    // Duration - Numeric returns Duration
-    if let Some(a) = instance_duration_value(&left)
-        && right.is_numeric()
-    {
-        let delta = runtime::to_float_value(&right).unwrap_or(0.0);
-        return make_duration(a - delta);
-    }
-    // DateTime - DateTime => Duration
-    if let (Some((ly, lm, ld, lh, lmin, ls, ltz)), Some((ry, rm, rd, rh, rmin, rs, rtz))) = (
-        instance_datetime_parts(&left),
-        instance_datetime_parts(&right),
-    ) {
-        use crate::builtins::methods_0arg::temporal;
-        let left_instant = temporal::datetime_to_instant_leap_aware(ly, lm, ld, lh, lmin, ls, ltz);
-        let right_instant = temporal::datetime_to_instant_leap_aware(ry, rm, rd, rh, rmin, rs, rtz);
-        let secs = ((left_instant - right_instant) * 1_000_000.0).round() / 1_000_000.0;
-        return make_duration(secs);
-    }
-    // DateTime - Duration => DateTime
-    if let Some((y, m, d, h, mi, s, tz)) = instance_datetime_parts(&left)
-        && let Some(delta) = instance_duration_value(&right)
-    {
-        use crate::builtins::methods_0arg::temporal;
-        let instant = temporal::datetime_to_instant_leap_aware(y, m, d, h, mi, s, tz);
-        let (ny, nm, nd, nh, nmi, ns) =
-            temporal::instant_to_datetime_leap_aware(instant - delta, tz);
-        return temporal::make_datetime(ny, nm, nd, nh, nmi, ns, tz);
-    }
-    if let (Some(a), Some(b)) = (instance_days(&left), instance_days(&right)) {
-        return Value::Int(a - b);
-    }
-    if let Some(days) = instance_days(&left)
-        && let Value::Int(delta) = right
-    {
-        use crate::builtins::methods_0arg::temporal;
-        let new_days = days - delta;
-        let (y, m, d) = temporal::epoch_days_to_civil(new_days);
-        return temporal::make_date(y, m, d);
-    }
-    // Mixin-wrapped Range - Real: perform Range arithmetic and re-wrap
-    if let Some(result) = mixin_range_arith_val(left.clone(), right.clone(), arith_sub) {
-        return result;
-    }
-    // Range - Real: shift both bounds (only when the Range is on the left;
-    // `Real - Range` numifies the Range, matching Raku).
-    if let Some(range) = range_offset(&left, &right, arith_sub) {
-        return range;
-    }
-    let (l, r) = runtime::coerce_numeric(left, right);
-    if matches!(l, Value::Complex(_, _)) || matches!(r, Value::Complex(_, _)) {
-        let (ar, ai) = runtime::to_complex_parts(&l).unwrap_or((0.0, 0.0));
-        let (br, bi) = runtime::to_complex_parts(&r).unwrap_or((0.0, 0.0));
-        Value::Complex(ar - br, ai - bi)
-    } else if (matches!(l, Value::BigInt(_)) || matches!(r, Value::BigInt(_)))
-        && let (Some(a), Some(b)) = (as_bigint(&l), as_bigint(&r))
-    {
-        Value::from_bigint(a - b)
-    } else if let (Some((an, ad)), Some((bn, bd))) = (to_big_rat_parts(&l), to_big_rat_parts(&r))
-        && needs_bigrat_path(&l, &r)
-    {
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        if has_fat_rat {
-            match make_big_fat_rat(an * bd.clone() - bn * ad.clone(), ad * bd) {
-                Value::Rat(n, d) => Value::FatRat(n, d),
-                other => other,
-            }
-        } else {
-            make_big_rat_arith(an * bd.clone() - bn * ad.clone(), ad * bd)
-        }
-    } else if let (Some((an, ad)), Some((bn, bd))) =
-        (runtime::to_rat_parts(&l), runtime::to_rat_parts(&r))
-    {
-        let has_rat = matches!(l, Value::Rat(_, _) | Value::FatRat(_, _))
-            || matches!(r, Value::Rat(_, _) | Value::FatRat(_, _));
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        if has_rat {
-            if has_fat_rat {
-                let n = an * bd - bn * ad;
-                let d = ad * bd;
-                make_fat_rat(n, d)
-            } else {
-                rat_sub_checked(an, ad, bn, bd)
-            }
-        } else {
-            match (l, r) {
-                (Value::Int(a), Value::Int(b)) => match a.checked_sub(b) {
-                    Some(diff) => Value::Int(diff),
-                    None => Value::from_bigint(
-                        num_bigint::BigInt::from(a) - num_bigint::BigInt::from(b),
-                    ),
-                },
-                (Value::Num(a), Value::Num(b)) => Value::Num(a - b),
-                (Value::Int(a), Value::Num(b)) => Value::Num(a as f64 - b),
-                (Value::Num(a), Value::Int(b)) => Value::Num(a - b as f64),
-                _ => Value::Int(0),
-            }
-        }
-    } else if (matches!(l, Value::BigInt(_)) || matches!(r, Value::BigInt(_)))
-        && let (Some(a), Some(b)) = (as_bigint(&l), as_bigint(&r))
-    {
-        Value::from_bigint(a - b)
-    } else {
-        let lf = runtime::to_float_value(&l);
-        let rf = runtime::to_float_value(&r);
-        if let (Some(a), Some(b)) = (lf, rf) {
-            Value::Num(a - b)
-        } else {
-            match (l, r) {
-                (Value::Int(a), Value::Int(b)) => match a.checked_sub(b) {
-                    Some(diff) => Value::Int(diff),
-                    None => Value::from_bigint(
-                        num_bigint::BigInt::from(a) - num_bigint::BigInt::from(b),
-                    ),
-                },
-                (Value::Num(a), Value::Num(b)) => Value::Num(a - b),
-                (Value::Int(a), Value::Num(b)) => Value::Num(a as f64 - b),
-                (Value::Num(a), Value::Int(b)) => Value::Num(a - b as f64),
-                _ => Value::Int(0),
-            }
-        }
-    }
-}
-
-pub(crate) fn arith_mul(left: Value, right: Value) -> Value {
-    let (left, right) = (left.into_deref(), right.into_deref());
-    // Mixin-wrapped Range * Real: perform Range arithmetic and re-wrap
-    if let Some(result) = mixin_range_arith_val(left.clone(), right.clone(), arith_mul)
-        .or_else(|| mixin_range_arith_val(right.clone(), left.clone(), arith_mul))
-    {
-        return result;
-    }
-    // Range * Numeric: scale both bounds, preserving exclusivity
-    if let Some(range) = range_scale(&left, &right).or_else(|| range_scale(&right, &left)) {
-        return range;
-    }
-    let (l, r) = runtime::coerce_numeric(left, right);
-    if matches!(l, Value::Complex(_, _)) || matches!(r, Value::Complex(_, _)) {
-        let (ar, ai) = runtime::to_complex_parts(&l).unwrap_or((0.0, 0.0));
-        let (br, bi) = runtime::to_complex_parts(&r).unwrap_or((0.0, 0.0));
-        Value::Complex(ar * br - ai * bi, ar * bi + ai * br)
-    } else if (matches!(l, Value::BigInt(_)) || matches!(r, Value::BigInt(_)))
-        && let (Some(a), Some(b)) = (as_bigint(&l), as_bigint(&r))
-    {
-        Value::from_bigint(a * b)
-    } else if let (Some((an, ad)), Some((bn, bd))) = (to_big_rat_parts(&l), to_big_rat_parts(&r))
-        && needs_bigrat_path(&l, &r)
-    {
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        if has_fat_rat {
-            match make_big_fat_rat(an * bn, ad * bd) {
-                Value::Rat(n, d) => Value::FatRat(n, d),
-                other => other,
-            }
-        } else {
-            make_big_rat_arith(an * bn, ad * bd)
-        }
-    } else if let (Some((an, ad)), Some((bn, bd))) =
-        (runtime::to_rat_parts(&l), runtime::to_rat_parts(&r))
-    {
-        let has_rat = matches!(l, Value::Rat(_, _) | Value::FatRat(_, _))
-            || matches!(r, Value::Rat(_, _) | Value::FatRat(_, _));
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        if has_rat {
-            if has_fat_rat {
-                let n = an * bn;
-                let d = ad * bd;
-                make_fat_rat(n, d)
-            } else {
-                rat_mul_checked(an, ad, bn, bd)
-            }
-        } else {
-            match (l, r) {
-                (Value::Int(a), Value::Int(b)) => {
-                    if let Some(product) = a.checked_mul(b) {
-                        Value::Int(product)
-                    } else {
-                        Value::bigint(NumBigInt::from(a) * NumBigInt::from(b))
-                    }
-                }
-                (Value::Num(a), Value::Num(b)) => Value::Num(a * b),
-                (Value::Int(a), Value::Num(b)) => Value::Num(a as f64 * b),
-                (Value::Num(a), Value::Int(b)) => Value::Num(a * b as f64),
-                _ => Value::Int(0),
-            }
-        }
-    } else {
-        // When mixing Num with Rat, convert to float
-        let lf = runtime::to_float_value(&l);
-        let rf = runtime::to_float_value(&r);
-        if let (Some(a), Some(b)) = (lf, rf) {
-            Value::Num(a * b)
-        } else {
-            match (l, r) {
-                (Value::Int(a), Value::Int(b)) => {
-                    if let Some(product) = a.checked_mul(b) {
-                        Value::Int(product)
-                    } else {
-                        Value::bigint(NumBigInt::from(a) * NumBigInt::from(b))
-                    }
-                }
-                (Value::Num(a), Value::Num(b)) => Value::Num(a * b),
-                (Value::Int(a), Value::Num(b)) => Value::Num(a as f64 * b),
-                (Value::Num(a), Value::Int(b)) => Value::Num(a * b as f64),
-                _ => Value::Int(0),
-            }
-        }
-    }
-}
-
-pub(crate) fn arith_div(left: Value, right: Value) -> Result<Value, RuntimeError> {
-    let (left, right) = (left.into_deref(), right.into_deref());
-    // Mixin-wrapped Range / Real: perform Range arithmetic and re-wrap
-    if let Some(result) = mixin_range_arith(left.clone(), right.clone(), arith_div) {
-        return result;
-    }
-    // Range / Numeric: scale both bounds down
-    if let Some(range) = range_divide(&left, &right) {
-        return Ok(range);
-    }
-    let (l, r) = runtime::coerce_numeric(left, right);
-    if matches!(l, Value::Complex(_, _)) || matches!(r, Value::Complex(_, _)) {
-        let (ar, ai) = runtime::to_complex_parts(&l).unwrap_or((0.0, 0.0));
-        let (br, bi) = runtime::to_complex_parts(&r).unwrap_or((0.0, 0.0));
-        let denom = br * br + bi * bi;
-        if denom == 0.0 {
-            return Err(RuntimeError::numeric_divide_by_zero());
-        }
-        Ok(Value::Complex(
-            (ar * br + ai * bi) / denom,
-            (ai * br - ar * bi) / denom,
-        ))
-    } else {
-        // When mixing Num with Rat/FatRat, convert to float
-        let has_num = matches!(l, Value::Num(_)) || matches!(r, Value::Num(_));
-        let has_rat = matches!(l, Value::Rat(_, _) | Value::FatRat(_, _))
-            || matches!(r, Value::Rat(_, _) | Value::FatRat(_, _));
-        if has_num && has_rat {
-            let lf = runtime::to_float_value(&l).unwrap_or(0.0);
-            let rf = runtime::to_float_value(&r).unwrap_or(1.0);
-            if rf == 0.0 {
-                return Ok(RuntimeError::divide_by_zero_failure(None, None));
-            }
-            return Ok(Value::Num(lf / rf));
-        }
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        let has_big_rat = matches!(l, Value::BigRat(_, _)) || matches!(r, Value::BigRat(_, _));
-        if (has_fat_rat || has_big_rat)
-            && let (Some((an, ad)), Some((bn, bd))) = (to_big_rat_parts(&l), to_big_rat_parts(&r))
-        {
-            let result = if has_fat_rat {
-                make_big_fat_rat(an * bd, ad * bn)
-            } else {
-                make_big_rat_arith(an * bd, ad * bn)
-            };
-            return Ok(match result {
-                Value::Rat(n, d) if has_fat_rat => Value::FatRat(n, d),
-                other => other,
-            });
-        }
-        if (matches!(l, Value::BigInt(_)) || matches!(r, Value::BigInt(_)))
-            && let (Some(a), Some(b)) = (as_bigint(&l), as_bigint(&r))
-        {
-            if b.is_zero() {
-                return Err(RuntimeError::numeric_divide_by_zero());
-            }
-            // `/` on integers always yields a Rat, even when it divides evenly:
-            // `555…555 / 5` is `Rat 111…111/1`, not an Int. Collapsing an
-            // exact big quotient to a BigInt here lost the Rat-ness (its
-            // `.WHAT` became `Int` and `.raku` dropped the trailing `.0`).
-            return Ok(make_big_rat_arith(a, b));
-        }
-        if let (Some((an, ad)), Some((bn, bd))) =
-            (runtime::to_rat_parts(&l), runtime::to_rat_parts(&r))
-        {
-            let has_rat = matches!(l, Value::Rat(_, _) | Value::FatRat(_, _))
-                || matches!(r, Value::Rat(_, _) | Value::FatRat(_, _));
-            let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-            if has_rat || matches!((&l, &r), (Value::Int(_), Value::Int(_))) {
-                return Ok(if has_fat_rat {
-                    if let (Some(n), Some(d)) = (an.checked_mul(bd), ad.checked_mul(bn)) {
-                        make_fat_rat(n, d)
-                    } else {
-                        let n = NumBigInt::from(an) * NumBigInt::from(bd);
-                        let d = NumBigInt::from(ad) * NumBigInt::from(bn);
-                        let result = make_big_fat_rat(n, d);
-                        match result {
-                            Value::Rat(n, d) => Value::FatRat(n, d),
-                            other => other,
-                        }
-                    }
-                } else {
-                    rat_div_checked(an, ad, bn, bd)
-                });
-            }
-        }
-        Ok(match (&l, &r) {
-            (Value::Num(a), Value::Num(b)) if *b == 0.0 => {
-                return Ok(RuntimeError::divide_by_zero_failure(
-                    Some(Value::Num(*a)),
-                    Some("/"),
-                ));
-            }
-            (Value::Int(a), Value::Num(b)) if *b == 0.0 => {
-                return Ok(RuntimeError::divide_by_zero_failure(
-                    Some(Value::Int(*a)),
-                    Some("/"),
-                ));
-            }
-            (Value::Num(_), Value::Int(b)) if *b == 0 => {
-                return Ok(RuntimeError::divide_by_zero_failure(
-                    Some(l.clone()),
-                    Some("/"),
-                ));
-            }
-            (Value::Num(a), Value::Num(b)) => Value::Num(a / b),
-            (Value::Int(a), Value::Num(b)) => Value::Num(*a as f64 / b),
-            (Value::Num(a), Value::Int(b)) => Value::Num(a / *b as f64),
-            (Value::Num(a), Value::BigInt(b)) => {
-                let bf = runtime::to_float_value(&Value::BigInt(b.clone())).unwrap_or(1.0);
-                if bf == 0.0 {
-                    return Ok(RuntimeError::divide_by_zero_failure(
-                        Some(Value::Num(*a)),
-                        Some("/"),
-                    ));
-                }
-                Value::Num(a / bf)
-            }
-            (Value::BigInt(a), Value::Num(b)) => {
-                if *b == 0.0 {
-                    return Ok(RuntimeError::divide_by_zero_failure(
-                        Some(Value::BigInt(a.clone())),
-                        Some("/"),
-                    ));
-                }
-                let af = runtime::to_float_value(&Value::BigInt(a.clone())).unwrap_or(0.0);
-                Value::Num(af / b)
-            }
-            _ => Value::Int(0),
-        })
-    }
-}
-
-pub(crate) fn arith_mod(left: Value, right: Value) -> Result<Value, RuntimeError> {
-    let (left, right) = (left.into_deref(), right.into_deref());
-    if let Some(raw) = instance_duration_raw_value(&left)
-        && right.is_numeric()
-    {
-        // Duration % Real => Duration. Compute with exact rational arithmetic
-        // (the same code path as a plain modulo) so the result is eqv to
-        // Duration.new($seconds % $real). arith_mod throws X::Numeric::DivideByZero
-        // when the divisor is zero.
-        let modded = arith_mod(real_to_rat(&raw), right)?;
-        return Ok(make_duration_from_value(real_to_rat(&modded)));
-    }
-    let (mut l, mut r) = runtime::coerce_numeric(left, right);
-    // Mixed Num/Rat modulo should use floating semantics; routing through
-    // exact-rational reduction loses expected precision behavior for cases like
-    // 1.01 % 0.2 (should be ~0.01).
-    if matches!(l, Value::Num(_))
-        && matches!(
-            r,
-            Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-        )
-    {
-        r = Value::Num(runtime::to_float_value(&r).unwrap_or(f64::NAN));
-    } else if matches!(r, Value::Num(_))
-        && matches!(
-            l,
-            Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-        )
-    {
-        l = Value::Num(runtime::to_float_value(&l).unwrap_or(f64::NAN));
-    }
-    // Exact rational modulo (divisor-sign semantics) via big-rational parts so
-    // that large operands such as `(7/1) % (2**66)` are handled exactly instead
-    // of falling through to a 0 result.
-    let l_is_rat = matches!(
-        l,
-        Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-    );
-    let r_is_rat = matches!(
-        r,
-        Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-    );
-    if (l_is_rat || r_is_rat)
-        && let (Some((an, ad)), Some((bn, bd))) = (to_big_rat_parts(&l), to_big_rat_parts(&r))
-    {
-        if bn.is_zero() {
-            return Err(RuntimeError::numeric_divide_by_zero());
-        }
-        let num = num_integer::Integer::mod_floor(&(&an * &bd), &(&ad * &bn));
-        let den = &ad * &bd;
-        let has_fat_rat = is_fat_rat_like(&l) || is_fat_rat_like(&r);
-        return Ok(if has_fat_rat {
-            // FatRat % ... stays a FatRat. make_big_fat_rat normalizes small
-            // results down to Rat, so re-tag those as FatRat.
-            match crate::value::make_big_fat_rat(num, den) {
-                Value::Rat(n, d) => Value::FatRat(n, d),
-                other => other,
-            }
-        } else {
-            crate::value::make_big_rat_arith(num, den)
-        });
-    }
-    {
-        // Integer `%` by zero reports the dividend and `using %`, like Rakudo
-        // (`Attempt to divide 7 by zero using %`).
-        let mod_div0 = |dividend: &Value| {
-            RuntimeError::numeric_divide_by_zero_full(Some(dividend.clone()), Some("%"))
-        };
-        Ok(match (l, r) {
-            (Value::Int(a), Value::Int(0)) => {
-                return Err(mod_div0(&Value::Int(a)));
-            }
-            (Value::BigInt(a), Value::Int(0)) => {
-                return Err(mod_div0(&Value::from_bigint((*a).clone())));
-            }
-            (Value::Int(a), Value::BigInt(b)) if b.is_zero() => {
-                return Err(mod_div0(&Value::Int(a)));
-            }
-            (Value::BigInt(a), Value::BigInt(b)) if b.is_zero() => {
-                return Err(mod_div0(&Value::from_bigint((*a).clone())));
-            }
-            (Value::Int(a), Value::Int(b)) => Value::Int(num_integer::Integer::mod_floor(&a, &b)),
-            (Value::BigInt(a), Value::Int(b)) => {
-                let bb = num_bigint::BigInt::from(b);
-                Value::from_bigint(num_integer::Integer::mod_floor(a.as_ref(), &bb))
-            }
-            (Value::Int(a), Value::BigInt(b)) => {
-                let aa = num_bigint::BigInt::from(a);
-                Value::from_bigint(num_integer::Integer::mod_floor(&aa, b.as_ref()))
-            }
-            (Value::BigInt(a), Value::BigInt(b)) => {
-                Value::from_bigint(num_integer::Integer::mod_floor(a.as_ref(), b.as_ref()))
-            }
-            (Value::Num(a), Value::Num(0.0)) => {
-                return Ok(RuntimeError::divide_by_zero_failure(
-                    Some(Value::Num(a)),
-                    Some("%"),
-                ));
-            }
-            (Value::Num(a), Value::Num(b)) => Value::Num(float_mod_floor(a, b)),
-            (ref lv, Value::Num(0.0)) => {
-                return Ok(RuntimeError::divide_by_zero_failure(
-                    Some(lv.clone()),
-                    Some("%"),
-                ));
-            }
-            (Value::Int(a), Value::Num(b)) => Value::Num(float_mod_floor(a as f64, b)),
-            (Value::Num(a), Value::Int(b)) => Value::Num(float_mod_floor(a, b as f64)),
-            _ => Value::Int(0),
-        })
-    }
-}
-
-pub(crate) fn arith_pow(left: Value, right: Value) -> Value {
-    let (left, right) = (left.into_deref(), right.into_deref());
-    let (l, r) = runtime::coerce_numeric(left, right);
-    let to_pow_f64 = |v: &Value| -> Option<f64> {
-        runtime::to_float_value(v).or_else(|| match v {
-            Value::BigRat(n, d) => {
-                if d.is_zero() {
-                    if n.is_zero() {
-                        Some(f64::NAN)
-                    } else if n.sign() == Sign::Minus {
-                        Some(f64::NEG_INFINITY)
-                    } else {
-                        Some(f64::INFINITY)
-                    }
-                } else {
-                    Some(n.to_f64()? / d.to_f64()?)
-                }
-            }
-            _ => None,
-        })
-    };
-    let powf_or_zero = |base: &Value, exp: &Value| -> Value {
-        match (to_pow_f64(base), to_pow_f64(exp)) {
-            (Some(a), Some(b)) => Value::Num(a.powf(b)),
-            _ => Value::Int(0),
-        }
-    };
-    if matches!(l, Value::Complex(_, _)) || matches!(r, Value::Complex(_, _)) {
-        let (ar, ai) = runtime::to_complex_parts(&l).unwrap_or((0.0, 0.0));
-        let (br, bi) = runtime::to_complex_parts(&r).unwrap_or((0.0, 0.0));
-        // Special case: 0 ** anything = 0+0i (except 0**0 = 1+0i)
-        if ar == 0.0 && ai == 0.0 {
-            return if br == 0.0 && bi == 0.0 {
-                Value::Complex(1.0, 0.0)
-            } else {
-                Value::Complex(0.0, 0.0)
-            };
-        }
-        // Integer real exponent: compute by repeated multiplication
-        // (exponentiation by squaring) so the result stays exact, matching
-        // raku. The polar `exp(b*ln a)` form below introduces floating-point
-        // noise (e.g. `(0+1i)**2` would be `-1+1.2e-16i` instead of `-1+0i`).
-        if bi == 0.0 && br.fract() == 0.0 && br.abs() <= 1024.0 {
-            let cmul = |xr: f64, xi: f64, yr: f64, yi: f64| -> (f64, f64) {
-                (xr * yr - xi * yi, xr * yi + xi * yr)
-            };
-            let mut e = br.abs() as u64;
-            let (mut rr, mut ri) = (1.0_f64, 0.0_f64);
-            let (mut pr, mut pi) = (ar, ai);
-            while e > 0 {
-                if e & 1 == 1 {
-                    let (nr, ni) = cmul(rr, ri, pr, pi);
-                    rr = nr;
-                    ri = ni;
-                }
-                let (sr, si) = cmul(pr, pi, pr, pi);
-                pr = sr;
-                pi = si;
-                e >>= 1;
-            }
-            if br < 0.0 {
-                // Reciprocal: 1 / (rr + ri·i) = (rr - ri·i) / (rr² + ri²).
-                let denom = rr * rr + ri * ri;
-                return Value::Complex(rr / denom, -ri / denom);
-            }
-            return Value::Complex(rr, ri);
-        }
-        let ln_r = (ar * ar + ai * ai).sqrt().ln();
-        let ln_i = ai.atan2(ar);
-        let wr = br * ln_r - bi * ln_i;
-        let wi = br * ln_i + bi * ln_r;
-        let mag = wr.exp();
-        Value::Complex(mag * wi.cos(), mag * wi.sin())
-    } else {
-        match (l, r) {
-            (Value::Int(a), Value::Int(b)) if b >= 0 => {
-                if let Some(result) = a.checked_pow(b as u32) {
-                    Value::Int(result)
-                } else {
-                    // Overflow: use BigInt
-                    use num_bigint::BigInt;
-                    let base = BigInt::from(a);
-                    Value::bigint(base.pow(b as u32))
-                }
-            }
-            (Value::Int(a), Value::Int(b)) => {
-                let pos = (-b) as u32;
-                if let Some(base) = a.checked_pow(pos) {
-                    make_rat(1, base)
-                } else {
-                    Value::Num(1.0 / (a as f64).powi(pos as i32))
-                }
-            }
-            (Value::Rat(n, d), Value::Int(b)) if b >= 0 => {
-                let p = b as u32;
-                if let (Some(np), Some(dp)) = (n.checked_pow(p), d.checked_pow(p)) {
-                    make_rat(np, dp)
-                } else {
-                    // Overflow in i64: use BigInt, then check if denom fits
-                    let nn = NumBigInt::from(n).pow(p);
-                    let dd = NumBigInt::from(d).pow(p);
-                    match make_big_rat_arith(nn, dd) {
-                        // If make_big_rat reduced it to fit in i64, keep as Rat
-                        Value::Rat(rn, rd) => Value::Rat(rn, rd),
-                        // BigRat means denom exceeds i64: degrade to Num
-                        Value::BigRat(bn, bd) => {
-                            // Check if denom fits in i64 (numerator can be big)
-                            if let Some(d64) = bd.to_i64() {
-                                Value::bigrat(*bn, NumBigInt::from(d64))
-                            } else {
-                                Value::Num(bigint_ratio_to_f64(bn.as_ref(), bd.as_ref()))
-                            }
-                        }
-                        other => other,
-                    }
-                }
-            }
-            (Value::Rat(n, d), Value::Int(b)) => {
-                let p = (-b) as u32;
-                if let (Some(dp), Some(np)) = (d.checked_pow(p), n.checked_pow(p)) {
-                    make_rat(dp, np)
-                } else {
-                    let nn = NumBigInt::from(d).pow(p);
-                    let dd = NumBigInt::from(n).pow(p);
-                    match make_big_rat_arith(nn, dd) {
-                        Value::Rat(rn, rd) => Value::Rat(rn, rd),
-                        Value::BigRat(bn, bd) => {
-                            if let Some(d64) = bd.to_i64() {
-                                Value::bigrat(*bn, NumBigInt::from(d64))
-                            } else {
-                                Value::Num(bigint_ratio_to_f64(bn.as_ref(), bd.as_ref()))
-                            }
-                        }
-                        other => other,
-                    }
-                }
-            }
-            (Value::FatRat(n, d), Value::Int(b)) if b >= 0 => {
-                let p = b as u32;
-                if let (Some(np), Some(dp)) = (n.checked_pow(p), d.checked_pow(p)) {
-                    make_fat_rat(np, dp)
-                } else {
-                    let nn = NumBigInt::from(n).pow(p);
-                    let dd = NumBigInt::from(d).pow(p);
-                    match make_big_fat_rat(nn, dd) {
-                        Value::Rat(nn, dd) => Value::FatRat(nn, dd),
-                        other => other,
-                    }
-                }
-            }
-            (Value::FatRat(n, d), Value::Int(b)) => {
-                let p = (-b) as u32;
-                if let (Some(dp), Some(np)) = (d.checked_pow(p), n.checked_pow(p)) {
-                    make_fat_rat(dp, np)
-                } else {
-                    let nn = NumBigInt::from(d).pow(p);
-                    let dd = NumBigInt::from(n).pow(p);
-                    match make_big_fat_rat(nn, dd) {
-                        Value::Rat(nn, dd) => Value::FatRat(nn, dd),
-                        other => other,
-                    }
-                }
-            }
-            (Value::BigRat(n, d), Value::Int(b)) if b >= 0 => {
-                let p = b as u32;
-                make_big_rat_arith(n.pow(p), d.pow(p))
-            }
-            (Value::BigRat(n, d), Value::Int(b)) => {
-                let p = (-b) as u32;
-                make_big_rat_arith(d.pow(p), n.pow(p))
-            }
-            (Value::Num(a), Value::Int(b)) => Value::Num(a.powi(b as i32)),
-            (Value::BigInt(a), Value::Int(b)) if b >= 0 => {
-                if let Ok(exp_u) = u32::try_from(b) {
-                    Value::bigint((*a).clone().pow(exp_u))
-                } else {
-                    Value::Num(a.as_ref().to_f64().unwrap_or(0.0).powf(b as f64))
-                }
-            }
-            (Value::BigInt(a), Value::Int(b)) => {
-                if *a == NumBigInt::from(1) {
-                    Value::Int(1)
-                } else if *a == NumBigInt::from(-1) {
-                    if (-b) % 2 == 0 {
-                        Value::Int(1)
-                    } else {
-                        Value::Int(-1)
-                    }
-                } else {
-                    Value::Num(a.as_ref().to_f64().unwrap_or(0.0).powf(b as f64))
-                }
-            }
-            (Value::Int(a), Value::Num(b)) => {
-                if b.is_finite() && b.fract() == 0.0 && b >= i32::MIN as f64 && b <= i32::MAX as f64
-                {
-                    let exp_i = b as i32;
-                    if exp_i >= 0 {
-                        let exp_u = exp_i as u32;
-                        if let Some(result) = a.checked_pow(exp_u) {
-                            Value::Int(result)
-                        } else {
-                            use num_bigint::BigInt;
-                            Value::bigint(BigInt::from(a).pow(exp_u))
-                        }
-                    } else {
-                        let pos = (-exp_i) as u32;
-                        if let Some(base) = a.checked_pow(pos) {
-                            make_rat(1, base)
-                        } else {
-                            Value::Num(1.0 / (a as f64).powi(pos as i32))
-                        }
-                    }
-                } else {
-                    Value::Num((a as f64).powf(b))
-                }
-            }
-            (Value::Int(a), Value::BigInt(b)) => {
-                let is_even = (b.as_ref() % 2u8) == NumBigInt::from(0u8);
-                if b.is_zero() || a == 1 {
-                    Value::Int(1)
-                } else if a == -1 {
-                    Value::Int(if is_even { 1 } else { -1 })
-                } else if a == 0 && b.sign() != Sign::Minus {
-                    Value::Int(0)
-                } else if let Some(exp_u) = b.as_ref().to_u32() {
-                    if let Some(result) = a.checked_pow(exp_u) {
-                        Value::Int(result)
-                    } else {
-                        Value::bigint(NumBigInt::from(a).pow(exp_u))
-                    }
-                } else {
-                    Value::Num((a as f64).powf(b.as_ref().to_f64().unwrap_or(f64::INFINITY)))
-                }
-            }
-            (Value::Num(a), Value::BigInt(b)) => {
-                let is_even = (b.as_ref() % 2u8) == NumBigInt::from(0u8);
-                if b.is_zero() || a == 1.0 {
-                    Value::Num(1.0)
-                } else if a == -1.0 {
-                    Value::Num(if is_even { 1.0 } else { -1.0 })
-                } else if a == 0.0 && b.sign() != Sign::Minus {
-                    Value::Num(0.0)
-                } else {
-                    Value::Num(a.powf(b.as_ref().to_f64().unwrap_or(f64::INFINITY)))
-                }
-            }
-            (Value::Num(a), Value::Num(b)) => {
-                let result = a.powf(b);
-                if result == 0.0 && a != 0.0 && b.is_finite() && b < 0.0 {
-                    RuntimeError::numeric_underflow_failure()
-                } else {
-                    Value::Num(result)
-                }
-            }
-            (base, exp) => powf_or_zero(&base, &exp),
-        }
-    }
-}
-
-pub(crate) fn arith_negate(val: Value) -> Result<Value, RuntimeError> {
-    match val {
-        Value::Bool(b) => Ok(Value::Int(if b { -1 } else { 0 })),
-        Value::Int(i) => {
-            if let Some(neg) = i.checked_neg() {
-                Ok(Value::Int(neg))
-            } else {
-                // i64::MIN overflow: promote to Num
-                Ok(Value::Num(-(i as f64)))
-            }
-        }
-        Value::BigInt(i) => Ok(Value::bigint(-(*i).clone())),
-        Value::Num(f) => Ok(Value::Num(-f)),
-        Value::Rat(n, d) => {
-            if let Some(neg) = n.checked_neg() {
-                Ok(Value::Rat(neg, d))
-            } else {
-                Ok(Value::Num(-(n as f64) / d as f64))
-            }
-        }
-        Value::FatRat(n, d) => {
-            if let Some(neg) = n.checked_neg() {
-                Ok(make_fat_rat(neg, d))
-            } else {
-                Ok(Value::Num(-(n as f64) / d as f64))
-            }
-        }
-        Value::BigRat(ref n, ref d) => {
-            if is_fat_rat_like(&val) {
-                Ok(make_big_fat_rat(-(**n).clone(), (**d).clone()))
-            } else {
-                Ok(make_big_rat_arith(-(**n).clone(), (**d).clone()))
-            }
-        }
-        Value::Complex(r, i) => {
-            // Canonicalize -0.0 to 0.0 so that e.g. (-i).reals gives (0e0, -1e0)
-            let nr = if r == 0.0 { 0.0 } else { -r };
-            let ni = if i == 0.0 { 0.0 } else { -i };
-            Ok(Value::Complex(nr, ni))
-        }
-        Value::Str(ref s) => {
-            if let Ok(i) = s.trim().parse::<i64>() {
-                Ok(Value::Int(-i))
-            } else if let Ok(f) = s.trim().parse::<f64>() {
-                Ok(Value::Num(-f))
-            } else {
-                Ok(Value::Int(0))
-            }
-        }
-        other => Ok(Value::Int(-runtime::to_int(&other))),
-    }
-}
+// Helpers used by external callers (vm/, runtime/).
+pub(crate) use rat::real_to_rat;
+pub(crate) use temporal::{is_temporal_operand, make_duration_value};
